test: trim fake-scaffolding tests, relocate prod-helper tests, de-repeat cms router
Precision cleanup of the test surface around the new @openora/core/testing fakes.

- delete the fake-behavior tests that only exercised the test doubles (InMemoryBroker
  fan-out, InProcess{Cache,RateLimiter} ttl/window, realtime transport/authorizer);
  they are already covered indirectly by every bootTestApp/domain test that uses them
- keep testing/fakes/__tests__/job-queue.test.ts - the fake carries real retry/backoff/
  dead-letter/idempotency/ordering logic domain tests don't exercise (safety net)
- relocate the production-helper tests (cached/invalidate, assertRateLimit/
  makeRateLimitError) out of the fakes dir into server/kernel/__tests__
- collapse cms.router's 6 identical authz rejects into one it.each table (same coverage)
- rename cms.cache.test.ts -> cms.reads.test.ts (2/3 tests were draft-filtering, not cache)

544 -> 523 tests, all green; no domain/security/money/compliance coverage removed.
